Importer changed to use updated data format: csv files, with multiple values in a single field delimited by the pipe symbol. Minor changes to error logging. The data model now uses floats for min/max pairs.
git-svn-id: http://svn.jazkarta.com/newfs-database-api/gobotany/trunk/src@345 fafe9cc9-101d-4c6e-bffa-87022359e7d1

+++ gobotany/core/models.py
@@ -185,8 +185,8 @@ class CharacterValue(models.Model):
     """
 
     value_str = models.CharField(max_length=260, null=True, blank=True)
-    value_min = models.IntegerField(null=True, blank=True)
-    value_max = models.IntegerField(null=True, blank=True)
+    value_min = models.FloatField(null=True, blank=True)
+    value_max = models.FloatField(null=True, blank=True)
     value_flt = models.FloatField(null=True, blank=True)
 
     character = models.ForeignKey(Character, related_name='character_values')
